refactor: extract useEditState hook, ColumnHeader component, and remove dead code
- Delete src/lib/pkce.js (unused PKCE helpers, implicit OAuth flow is used)
- Remove 3 unused mutation hooks from useContactPersons.js
- Extract ColumnHeader to src/components/ColumnHeader.jsx
- Extract useEditState hook (src/hooks/useEditState.js) with all dirtyMap,
  validationErrors, pendingPerson*, and localStorage persistence logic
- Import getContactFieldValue from csvImport.js into CustomerTable instead
  of duplicating the billing address / custom field resolution inline
- CustomerTable shrinks from 1,280 → 923 lines (28% reduction)

Closes #60
